### 2020-03-09 #21 @agbarnett
Ten Years Reproducibility Challenge: paper number #22
**Original article:** Barnett AG, Dobson AJ, For the WHO MONICA Project (2004) Estimating trends and seasonality in coronary heart disease. Statist Med 23:3505-3523. DOI: 10.1002/sim.1927

**Scientific domain:** Epidemiology
**Programming language:** SAS
