### 2020-04-14 #21 @rougier
Sorry I'm very late.
@ctb @benoit-girard @oliviaguest Can any of you edit this submission for the Ten Years Reproducibility Challenge (only 1 reviewer needed). I know it's not in your domain but the https://rescience.github.io/board/ might help.
