### 2020-06-15 #21 @rougier
@ctb @benoit-girard @oliviaguest @khinsen  (kind of urgent) Can any of you edit this submission for the Ten Years Reproducibility Challenge (only 1 reviewer needed). I know it's not in your domain but the https://rescience.github.io/board/ might help.
