### 2020-06-21 #21 @khinsen
@SergeStinckwich Can you review this reproducibility report in computational epidemiology? Or propose other reviewers?
