### 2020-06-22 #21 @SergeStinckwich
> @SergeStinckwich Can you review this reproducibility report in computational epidemiology? Or propose other reviewers?

I don't know SAS but I can help if needed. I will be free after June 28th to do a review.
