### 2020-06-22 #21 @khinsen
@SergeStinckwich Next week is fine! As for SAS, it would be sheer luck to find someone with that competence in the Open Source oriented circle of ReScience. There's a good reason we don't accept proprietary software in regular submissions. We made an exception for this challenge, so we have to find a compromise to deal with it.
