### 2020-11-23 #21 @SergeStinckwich
Sorry for the delay. I was able to spend sometimes to read the paper. The paper is well written, the author was able to reproduce the original results. Small differences in the reproduced results are explained clearly. As I don't have access to SAS, I'm not able to reproduce the result from my side. Minor issue: apparently the authors did not put the original code in the git repository.
