### 2020-11-24 #21 @SergeStinckwich
> Thanks. The original code is [here](https://github.com/agbarnett/tenyears) on github in the 'code' subfolder.

In fact, this was not clear for me, what was the difference between original code and new one. Thank you for clarification.
