### 2020-11-24 #21 @khinsen
@SergeStinckwich Do I understand correctly that in your opinion, this paper can be accepted? We'll have to live with not evaluating the SAS code.

My own impression of the article is that this is a nice (and very readable) story illustrating the importance of properly archiving and documenting  code and data.
