### 2020-11-25 #21 @khinsen
Thanks @SergeStinckwich for the clarification, and of course for the review!

Herewith I proclaim this paper ACCEPTED! I will do the publication steps later today or tomorrow.
