### 2020-11-25 #21 @khinsen
@agbarnett I submitted a [pull request](https://github.com/agbarnett/tenyears/pull/2) with the updated metadata for publication. But I cannot rebuild the article PDF myself because one figure (`figures/original_figure.eps`) is missing from your repository. Could you please update the PDF yourself using the new metadata file?
