### 2020-11-25 #21 @agbarnett
@khinsen I've added the missing file. Is python the only way to generate the metadata.tex file from metadata.yaml? I've never used python.
