### 2020-11-26 #21 @khinsen
Thanks, with the added file I can rebuild the article myself. The metadata is indeed processed with a Python script, but you don't need to know any Python to use it. All you need is Python 3 plus PyYAML (see https://github.com/rescience/template for instructions).

What I did to rebuild your article is:
- Clone a fresh copy of https://github.com/rescience/template
- Copy over `content.tex`, `bibliography.bib`, `metadata.yaml`, and `figures` from your `article` directory
- Run `make`

I just did another save request at Software Heritage because of the added figure file. I'll then update the SWHID in the metadata, rebuild the article, and submit another pull request with the new stuff. And publish the article on Zenodo of course.

Our software infrastructure is unfortunately still a bit rough. That's the price to pay for getting reproducible article PDFs with permalinks to code. One day mainstream publishing systems will catch up with us!
