### 2020-11-26 #21 @khinsen
The paper is published: http://doi.org/10.5281/zenodo.4290512
